fix(jsutils): Clean up noOps and jsDocs
