+++ repos/jsutils/src/array/flatUnion.js
@@ -1,3 +1,5 @@
+/** @module array */
+
 import { isArr } from './isArr'
 import { uniqArr } from './uniqArr'
 import { flatArr } from './flatArr'
