+++ repos/jsutils/src/node/findProc.js
@@ -1,3 +1,5 @@
+/** @module node */
+
 const { exec } = require('child_process')
 
 /**
