+++ repos/jsutils/src/node/inDocker.js
@@ -1,3 +1,5 @@
+/** @module node */
+
 const fs = require('fs')
 let inContainer
