+++ repos/jsutils/src/node/loadModule.js
@@ -1,4 +1,4 @@
-/** @module command */
+/** @module node */
 
 'use strict'
