+++ repos/jsutils/src/string/camelCasePath.js
@@ -1,3 +1,5 @@
+/** @module string */
+
 import { capitalize } from './capitalize'
 
 /**
